@@ -1,8 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>netstandard2.0;netstandard2.1;</TargetFrameworks>
-    <LangVersion>8.0</LangVersion>
+    <TargetFrameworks>netstandard2.0;netstandard2.1;Net6</TargetFrameworks>
+    <LangVersion>10.0</LangVersion>
     <nullable>enable</nullable>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
     <Description>Registration of Dispatching classes in a container from Grace.</Description>
@@ -15,13 +15,13 @@
     <Copyright>Copyright @2021 Olbrasoft</Copyright>
     <AssemblyName>Olbrasoft.Dispatching.DI.Grace.Common</AssemblyName>
     <RootNamespace>Olbrasoft.Dispatching.DI.Grace.Common</RootNamespace>
-    <Version>1.3.1</Version>
+    <Version>1.4.0</Version>
     <PackageIconUrl />
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="Grace" Version="7.2.1" />
-    <PackageReference Include="Olbrasoft.Dispatching.DI.Common" Version="1.3.0" />
+    <PackageReference Include="Olbrasoft.Dispatching.DI.Common" Version="1.4.0" />
   </ItemGroup>
 
   <ItemGroup>
