@@ -1,30 +1,33 @@
 <Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0;netstandard2.1;</TargetFrameworks>
-    <LangVersion>8.0</LangVersion>
-    <nullable>enable</nullable>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <Description>Registration of Dispatching classes in a container from Microsoft.</Description>
-    <PackageProjectUrl></PackageProjectUrl>
-    <PackageIcon>olbrasoft-dispatching.png</PackageIcon>
-    <RepositoryUrl>https://github.com/Olbrasoft/Dispatching</RepositoryUrl>
-    <PackageTags>Dependency Injection Mediator Dispatchig Command Query Responsibility Segredation</PackageTags>
-    <Authors>Jiří Tůma</Authors>
-    <Company>Olbrasoft</Company>
-    <Copyright>Copyright @2021 Olbrasoft</Copyright>
-    <AssemblyName>Olbrasoft.Dispatching.DI.Microsoft.Common</AssemblyName>
-    <RootNamespace>Olbrasoft.Dispatching.DI.Microsoft.Common</RootNamespace>
-    <Version>1.3.1</Version>
-    <PackageIconUrl />
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="5.0.0" />
-    <PackageReference Include="Olbrasoft.Dispatching.DI.Common" Version="1.3.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="..\..\olbrasoft-dispatching.png">
-      <Pack>True</Pack>
-      <PackagePath></PackagePath>
-    </None>
-  </ItemGroup>
+
+	<PropertyGroup>
+		<TargetFrameworks>Netstandard2.0;netstandard2.1;Net6</TargetFrameworks>
+		<LangVersion>10.0</LangVersion>
+		<nullable>enable</nullable>
+		<PackageLicenseExpression>MIT</PackageLicenseExpression>
+		<Description>Registration of Dispatching classes in a container from Microsoft.</Description>
+		<PackageProjectUrl></PackageProjectUrl>
+		<PackageIcon>olbrasoft-dispatching.png</PackageIcon>
+		<RepositoryUrl>https://github.com/Olbrasoft/Dispatching</RepositoryUrl>
+		<PackageTags>Dependency Injection Mediator Dispatchig Command Query Responsibility Segredation</PackageTags>
+		<Authors>Jiří Tůma</Authors>
+		<Company>Olbrasoft</Company>
+		<Copyright>Copyright @2021 Olbrasoft</Copyright>
+		<AssemblyName>Olbrasoft.Dispatching.DI.Microsoft.Common</AssemblyName>
+		<RootNamespace>Olbrasoft.Dispatching.DI.Microsoft.Common</RootNamespace>
+		<Version>1.4.0</Version>
+		<PackageIconUrl />
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="6.0.0" />
+		<PackageReference Include="Olbrasoft.Dispatching.DI.Common" Version="1.4.0" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<None Include="..\..\olbrasoft-dispatching.png">
+			<Pack>True</Pack>
+			<PackagePath></PackagePath>
+		</None>
+	</ItemGroup>
 </Project>
