@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>netstandard2.0;netstandard2.1;</TargetFrameworks>
-    <LangVersion>8.0</LangVersion>
+    <TargetFrameworks>netstandard2.0;netstandard2.1;Net6</TargetFrameworks>
+    <LangVersion>10.0</LangVersion>
     <nullable>enable</nullable>
     <PackageIcon>olbrasoft-dispatching.png</PackageIcon>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
@@ -9,7 +9,7 @@
     <PackageProjectUrl>https://github.com/Olbrasoft/Dispatching/</PackageProjectUrl>
     <PackageTags>Dependency Injection Mediator Dispatchig Command Query Responsibility Segredation</PackageTags>
     <Description>Registers classes for Dispatching with Executor dispatcher to the Microsoft container.</Description>
-    <Version>1.3.1</Version>
+    <Version>1.4.0</Version>
     <Authors>Jiří Tůma</Authors>
     <Company>Olbrasoft</Company>
     <Product>Dispatching</Product>
@@ -25,7 +25,7 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Olbrasoft.Dispatching" Version="1.2.0" />
-    <PackageReference Include="Olbrasoft.Dispatching.DI.Microsoft.Common" Version="1.2.0" />
+    <PackageReference Include="Olbrasoft.Dispatching" Version="1.4.0" />
+    <PackageReference Include="Olbrasoft.Dispatching.DI.Microsoft.Common" Version="1.4.0" />
   </ItemGroup>
 </Project>
