@@ -1,8 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>netstandard2.0;netstandard2.1;</TargetFrameworks>
-    <LangVersion>8.0</LangVersion>
+    <TargetFrameworks>netstandard2.0;netstandard2.1;Net6</TargetFrameworks>
+    <LangVersion>10.0</LangVersion>
     <nullable>enable</nullable>
     <Authors>Jiří Tůma</Authors>
     <Company>Olbrasoft</Company>
@@ -15,15 +15,15 @@
     <PackageIcon>olbrasoft-dispatching.png</PackageIcon>
     <RepositoryType>git</RepositoryType>
     <PackageTags>Dispatching Mediatr CQRS Dependency Injection Grace</PackageTags>
-    <Version>1.3.1</Version>
+    <Version>1.4.0</Version>
     <AssemblyName>Olbrasoft.Dispatching.Dynamic.DI.Grace</AssemblyName>
     <RootNamespace>Olbrasoft.Extensions.DependencyInjection</RootNamespace>
     <PackageIconUrl />
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Olbrasoft.Dispatching.DI.Grace.Common" Version="1.3.1" />
-    <PackageReference Include="Olbrasoft.Dispatching.Dynamic" Version="1.3.1" />
+    <PackageReference Include="Olbrasoft.Dispatching.DI.Grace.Common" Version="1.4.0" />
+    <PackageReference Include="Olbrasoft.Dispatching.Dynamic" Version="1.4.0" />
   </ItemGroup>
 
   <ItemGroup>
