@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>netstandard2.0;netstandard2.1;netstandard1.6;</TargetFrameworks>
-    <LangVersion>8.0</LangVersion>
+    <TargetFrameworks>netstandard2.0;netstandard2.1;netstandard1.6;Net6</TargetFrameworks>
+    <LangVersion>10.0</LangVersion>
     <nullable>enable</nullable>
     <Description>Dispatcher implemented using dynamics</Description>
     <PackageProjectUrl>https://github.com/Olbrasoft/Dispatching</PackageProjectUrl>
@@ -13,7 +13,7 @@
     <Company>Olbrasoft</Company>
     <Product>Dispatching</Product>
     <Copyright>Copyright ©2021 Olbrasoft</Copyright>
-    <Version>1.3.1</Version>
+    <Version>1.4.0</Version>
     <AssemblyName>Olbrasoft.Dispatching.Dynamic</AssemblyName>
     <RootNamespace>Olbrasoft.Dispatching</RootNamespace>
     <PackageId>Olbrasoft.Dispatching.Dynamic</PackageId>
@@ -27,6 +27,6 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
-    <PackageReference Include="Olbrasoft.Dispatching.Abstractions" Version="1.3.0" />
+    <PackageReference Include="Olbrasoft.Dispatching.Abstractions" Version="1.4.0" />
   </ItemGroup>
 </Project>
