@@ -1,30 +1,33 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0;netstandard2.1;netstandard1.6;</TargetFrameworks>
-    <LangVersion>8.0</LangVersion>
-    <nullable>enable</nullable>
-    <AssemblyName>Olbrasoft.Dispatching</AssemblyName>
-    <RootNamespace>Olbrasoft.Dispatching</RootNamespace>
-    <Copyright>Copyright @2021 Olbrasoft</Copyright>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <Product>Dispatching</Product>
-    <RepositoryUrl>https://github.com/Olbrasoft/Dispatching</RepositoryUrl>
-    <PackageLicenseExpression />
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageIcon>olbrasoft-dispatching.png</PackageIcon>
-    <PackageIconUrl />
-    <Version>1.3.0</Version>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFrameworks>netstandard2.0;netstandard2.1;netstandard1.6;Net6.0</TargetFrameworks>
+		<LangVersion>10.0</LangVersion>
+		<nullable>enable</nullable>
+		<AssemblyName>Olbrasoft.Dispatching</AssemblyName>
+		<RootNamespace>Olbrasoft.Dispatching</RootNamespace>
+		<Product>Dispatching</Product>
+		<RepositoryUrl>https://github.com/Olbrasoft/Dispatching</RepositoryUrl>
+		<PackageLicenseExpression />
+		<PackageLicenseExpression>MIT</PackageLicenseExpression>
+		<PackageIcon>olbrasoft-dispatching.png</PackageIcon>
+		<PackageIconUrl />
+		<Version>1.4.0</Version>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <None Include="..\..\olbrasoft-dispatching.png">
-      <Pack>True</Pack>
-      <PackagePath></PackagePath>
-    </None>
-  </ItemGroup>
+	<ItemGroup>
+		<None Include="..\..\olbrasoft-dispatching.png">
+			<Pack>True</Pack>
+			<PackagePath></PackagePath>
+		</None>
+	</ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Olbrasoft.Dispatching.Abstractions" Version="1.3.0" />
-  </ItemGroup>
+	<ItemGroup>
+		<Using Include="System.Threading"/>
+		<Using Include="System.Threading.Tasks"/>
+	</ItemGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Olbrasoft.Dispatching.Abstractions" Version="1.4.0" />
+	</ItemGroup>
 </Project>
