@@ -1,8 +1,7 @@
-namespace Olbrasoft.Dispatching
-{
-    /// <summary>
-    /// Marker interface to represent a request with a response
-    /// </summary>
-    /// <typeparam name="TResponse">Response type</typeparam>
-    public interface IRequest<out TResponse> { }
-}
+namespace Olbrasoft.Dispatching;
+
+/// <summary>
+/// Marker interface to represent a request with a response
+/// </summary>
+/// <typeparam name="TResponse">Response type</typeparam>
+public interface IRequest<out TResponse> { }
