@@ -1,6 +1,4 @@
-﻿using Olbrasoft.Dispatching.Abstractions;
-
-namespace Olbrasoft.Dispatching.Benchmarks
+﻿namespace Olbrasoft.Dispatching.Benchmarks
 {
     public class AwesomeRequest : IRequest<object>, MediatR.IRequest<object>
     {
