@@ -1,6 +1,4 @@
-﻿using Olbrasoft.Dispatching.Abstractions;
-
-namespace Olbrasoft.Dispatching.Benchmarks
+﻿namespace Olbrasoft.Dispatching.Benchmarks
 {
     public interface IDispatcherWithExecutor : IDispatcher
     {
