@@ -1,6 +1,4 @@
-﻿using Olbrasoft.Dispatching.Abstractions;
-
-namespace Olbrasoft.Dispatching.DI.Microsoft
+﻿namespace Olbrasoft.Dispatching.DI.Microsoft
 {
     public class AwesomeRequest : Request<object>
     {
